rename channel doctor and patient to caller and recipient
