+++ frontend/src/components/MeetingCallerView.css
@@ -1,0 +1,88 @@
+/*
+ * TODO: This styling needs to be adjusted based on where the existing widget is loaded.
+ * The audio video meeting widget needs to calculate the place that it can fit most accurately and then accordingly set the top and bottom properties.
+ */
+
+.MeetingCallerView__window {
+  left: calc(var(--chime-sdk-widget-width) + 2rem);
+}
+
+.MeetingCallerView {
+  align-items: center;
+  background-color: white;
+  display: flex;
+  flex-direction: column;
+  height: var(--chime-sdk-widget-content-height);
+  justify-content: center;
+}
+
+.MeetingCallerView__progressUpdateContainer {
+  align-items: center;
+  background-color: white;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  justify-content: center;
+  padding: 1rem;
+  text-align: center;
+  width: 100%;
+}
+
+.MeetingCallerView__waitingContainer {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  flex: 1 1 auto;
+  justify-content: center;
+}
+
+.MeetingCallerView__waiting {
+  margin: 0;
+}
+
+.MeetingCallerView__cancelButton {
+  margin-top: 0.25rem;
+  background-color: transparent;
+  border: none;
+  color: var(--chime-sdk-widget-button-color);
+  cursor: pointer;
+  transition: all 0.25s;
+  padding: 0.5rem 0.75rem;
+  border-radius: 0.25rem;
+}
+
+.MeetingCallerView__cancelButton:active,
+.MeetingCallerView__cancelButton:hover,
+.MeetingCallerView__cancelButton:focus {
+  color: var(--chime-sdk-widget-button-hover-color);
+  background-color: var(--chime-sdk-widget-link-hover-background-color);
+}
+
+.MeetingCallerView__phoneCallButton {
+  background-color: var(--chime-sdk-widget-button-color);
+  border-radius: 0.25rem;
+  border: none;
+  color: white;
+  cursor: pointer;
+  font-weight: 700;
+  padding: 0.5rem;
+  transition: all 0.25s;
+  width: 100%;
+}
+
+.MeetingCallerView__phoneCallButton:active,
+.MeetingCallerView__phoneCallButton:hover,
+.MeetingCallerView__phoneCallButton:focus {
+  background-color: var(--chime-sdk-widget-button-hover-color);
+}
+
+.MeetingCallerView__phoneCallButton[disabled] {
+  cursor: not-allowed;
+  color: var(--chime-sdk-widget-button-disabled-color);
+  background-color: var(--chime-sdk-widget-button-disabled-background-color);
+}
+
+.MeetingCallerView__phoneCallDescription {
+  font-size: 0.8rem;
+  color: var(--chime-sdk-widget-light-text-color);
+}
