+++ frontend/src/components/MeetingRecipientView.tsx
@@ -24,8 +24,7 @@ interface Props {
   onCleanUp: () => void;
 }
 
-export default function MeetingPatientView({
-  number,
+export default function MeetingRecipientView({
   channel,
   meetingId,
   onCleanUp,
@@ -109,11 +108,10 @@ export default function MeetingPatientView({
 
   return (
     <Window
-      number={number}
       className="MeetingPatientView__window"
       isPortal
       title={t('MeetingPatientView.title', {
-        name: channel.doctor.name,
+        name: channel.caller.name,
       })}
     >
       <div className="MeetingPatientView">
@@ -123,7 +121,7 @@ export default function MeetingPatientView({
               <Trans
                 i18nKey={'MeetingPatientView.received'}
                 values={{
-                  name: channel.doctor.name,
+                  name: channel.caller.name,
                 }}
               />
             </p>
@@ -147,7 +145,7 @@ export default function MeetingPatientView({
               <Trans
                 i18nKey={'MeetingPatientView.starting'}
                 values={{
-                  name: channel.doctor.name,
+                  name: channel.caller.name,
                 }}
               />
             </p>
@@ -159,7 +157,7 @@ export default function MeetingPatientView({
               attendee={joinInfo.Attendee}
               meeting={joinInfo.Meeting}
               onCleanUp={onCleanUp}
-              remoteAttendeeName={channel.doctor.name}
+              remoteAttendeeName={channel.caller.name}
             />
           </MeetingProvider>
         )}
